Update hello_world.py
fixed ermove task bug causing app to crash

+++ hello_world.py
@@ -37,6 +37,7 @@ import sys
 from PyQt6 import QtWidgets, uic, QtCore
 import sqlite3
 from PyQt6.QtCore import QResource
+import resources_rc
 
 # images contained in the resources.qrc file
 QResource.registerResource("resources.qrc")
@@ -68,6 +69,7 @@ date_input.setDate(QtCore.QDate.currentDate())
 # adding the todos from the database to the list view
 cursor.execute("SELECT * FROM todos")
 rows = cursor.fetchall()
+
 for row in rows:
     list_view.addItem(f"{row[0]} - {row[1]}")
 
@@ -81,14 +83,23 @@ def add_todo():
     conn.commit()   
     date_input.setDate(QtCore.QDate.currentDate())
     todo_input.clear()
+    cursor.execute("SELECT * FROM todos")
+    global rows
+    rows = cursor.fetchall()
+    print(rows)
 
 
 def remove_todo():
     current_row = list_view.currentRow()
-    if current_row != -1:
+    global rows
+    if current_row != -1 and current_row < len(rows):
         cursor.execute("DELETE FROM todos WHERE date = ? AND todo = ?", (rows[current_row]))
         conn.commit()
         list_view.takeItem(current_row)
+        cursor.execute("SELECT * FROM todos")
+        rows = cursor.fetchall()
+        print(rows)
+
